requirements.txt
elevenlabs==1.50.2
openai==1.57.0
python-dotenv==1.0.1
streamlit==1.41.1
pydub==0.25.1